api/requirements.txt
APScheduler==3.8.1
backports.zoneinfo==0.2.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
cryptography==36.0.2
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-PyMongo==2.3.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
ndg-httpsclient==0.5.1
numpy==1.22.1
pandas==1.3.5
pusher==3.2.0
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.3.0
pymongo==4.0.1
PyNaCl==1.5.0
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
threadpoolctl==3.1.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
Werkzeug==2.0.2
